% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_annotations.R
\name{add_annotations}
\alias{add_annotations}
\title{Add branded plot annotations}
\usage{
add_annotations(
  type,
  text = NA,
  text_x_position = NA,
  text_y_position = NA,
  segment_x_position_start = NA,
  segment_x_position_end = NA,
  segment_y_position_start = NA,
  segment_y_position_end = NA
)
}
\arguments{
\item{type}{Either \code{"text"} or \code{"segment"}.}

\item{text}{Annotation text for \code{type = "text"}.}

\item{text_x_position}{X position for text annotation.}

\item{text_y_position}{Y position for text annotation.}

\item{segment_x_position_start}{Starting x position for a segment annotation.}

\item{segment_x_position_end}{Ending x position for a segment annotation.}

\item{segment_y_position_start}{Starting y position for a segment annotation.}

\item{segment_y_position_end}{Ending y position for a segment annotation.}
}
\value{
A ggplot annotation layer.
}
\description{
Add branded plot annotations
}
