% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_grid_lines.R
\name{add_grid_lines}
\alias{add_grid_lines}
\title{Add major grid lines to a plot}
\usage{
add_grid_lines(horizontal = FALSE, vertical = FALSE)
}
\arguments{
\item{horizontal}{Add horizontal major grid lines.}

\item{vertical}{Add vertical major grid lines.}
}
\value{
A ggplot theme modification.
}
\description{
Add major grid lines to a plot
}
