% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_legend.R
\name{add_legend}
\alias{add_legend}
\title{Show a legend with CGD styling}
\usage{
add_legend(position, justification)
}
\arguments{
\item{position}{Legend position such as \code{"top"}, \code{"bottom"}, \code{"left"}, or \code{"right"}.}

\item{justification}{Legend justification passed to ggplot2.}
}
\value{
A ggplot theme modification.
}
\description{
Show a legend with CGD styling
}
