% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_projection_line.R
\name{add_projection_line}
\alias{add_projection_line}
\title{Add a dashed projection line}
\usage{
add_projection_line(df)
}
\arguments{
\item{df}{Data frame containing the projected portion of the series.}
}
\value{
A ggplot line layer.
}
\description{
Add a dashed projection line
}
