% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_separator_line.R
\name{add_separator_line}
\alias{add_separator_line}
\title{Add a separator line}
\usage{
add_separator_line(orientation = c("vertical", "horizontal"), intercept)
}
\arguments{
\item{orientation}{Either \code{"vertical"} or \code{"horizontal"}.}

\item{intercept}{Axis intercept where the line should be drawn.}
}
\value{
A ggplot line layer.
}
\description{
Add a separator line
}
