% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_trend_line.R
\name{add_trend_line}
\alias{add_trend_line}
\title{Add a dotted trend line}
\usage{
add_trend_line(method)
}
\arguments{
\item{method}{Smoothing method passed to \code{ggplot2::geom_smooth()}.}
}
\value{
A ggplot smooth layer.
}
\description{
Add a dotted trend line
}
