% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_cgd_palette.R
\name{cgd_palette}
\alias{cgd_palette}
\title{Return colors from a CGD palette}
\usage{
cgd_palette(palette_name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{palette_name}{Name of the palette to retrieve.}

\item{n}{Number of colors to return. Defaults to the full palette.}

\item{type}{Whether to return a discrete subset or a continuous interpolation.}
}
\value{
A character vector of hex color values.
}
\description{
Return colors from a CGD palette
}
