% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_pt.R
\name{convert_to_pt}
\alias{convert_to_pt}
\title{Convert a point size for ggplot geoms}
\usage{
convert_to_pt(size)
}
\arguments{
\item{size}{Size in typographic points.}
}
\value{
Numeric size in ggplot's internal units.
}
\description{
Convert a point size for ggplot geoms
}
