% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_cgd_palette.R
\name{load_cgd_palette}
\alias{load_cgd_palette}
\title{Return all packaged CGD palettes}
\usage{
load_cgd_palette()
}
\value{
A named list of palette vectors.
}
\description{
Return all packaged CGD palettes
}
