% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_annotations.R
\name{add_annotations}
\alias{add_annotations}
\title{A function}
\usage{
add_annotations(
  type,
  text = NA,
  text_x_position = NA,
  text_y_position = NA,
  segment_x_position_start = NA,
  segment_x_position_end = NA,
  segment_y_position_start = NA,
  segment_y_position_end = NA
)
}
\arguments{
\item{type}{xxx}

\item{text}{xxx}

\item{text_x_position}{xxx}

\item{text_y_position}{xxx}

\item{segment_x_position_start}{xxx}

\item{segment_x_position_end}{xxx}

\item{segment_y_position_start}{xxx}

\item{segment_y_position_end}{xxx}
}
\description{
A function
}
