% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_grid_lines.R
\name{add_grid_lines}
\alias{add_grid_lines}
\title{A function}
\usage{
add_grid_lines(horizontal = FALSE, vertical = FALSE)
}
\arguments{
\item{horizontal}{add horizontal grid lines T/F}

\item{vertical}{add vertical grid lines T/F}
}
\description{
A function
}
