% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_labels.R
\name{add_labels}
\alias{add_labels}
\title{Add data labels using CGD defaults}
\usage{
add_labels(label, chart_type)
}
\arguments{
\item{label}{Vector or column used for label text.}

\item{chart_type}{One of \code{"bar"}, \code{"stacked"}, or \code{"line"}.}
}
\value{
A ggplot layer for use with \code{+}.
}
\description{
Add data labels using CGD defaults
}
