% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_legend.R
\name{add_legend}
\alias{add_legend}
\title{A function}
\usage{
add_legend(position, justification)
}
\arguments{
\item{position}{top bottom left right of the chart}

\item{justification}{left right center}
}
\description{
A function
}
