% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_separator_line.R
\name{add_separator_line}
\alias{add_separator_line}
\title{A function}
\usage{
add_separator_line(orientation, intercept)
}
\arguments{
\item{orientation}{horizontal or vertical}

\item{intercept}{constant where the line will fall}
}
\description{
A function
}
