% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_cgd_palette.R
\name{cgd_palette}
\alias{cgd_palette}
\title{A function}
\usage{
cgd_palette(palette_name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{palette_name}{color palette of choice}

\item{n}{how many colors from the palette is required}

\item{type}{discrete or continuous}
}
\description{
A function
}
