% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_axis_text.R
\name{wrap_axis_text}
\alias{wrap_axis_text}
\title{A function}
\usage{
wrap_axis_text(add_space = FALSE, column, num_text = 0)
}
\arguments{
\item{add_space}{xxx}

\item{column}{yyy}

\item{num_text}{zzz}
}
\description{
A function
}
