% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_axis_text.R
\name{wrap_axis_text}
\alias{wrap_axis_text}
\title{A function}
\usage{
wrap_axis_text(add_space = FALSE, column, num_text)
}
\arguments{
\item{orientation}{horizontal or vertical}

\item{intercept}{constant where the line will fall}
}
\description{
A function
}
